Fix silent copies between python and c++ by using int64 isntead of int.

+++ src/cpp/_augment.cpp
@@ -57,13 +57,14 @@ namespace py = pybind11;
 #include <algorithm>
 #include <cmath>
 #include <vector>
+#include <cstdint>
 
 
 py::tuple
 augment(py::array_t<double> cost_matrix,
-        int cur_row,
-        py::array_t<int> row4col,
-        py::array_t<int> col4row,
+        int64_t cur_row,
+        py::array_t<int64_t> row4col,
+        py::array_t<int64_t> col4row,
         py::array_t<double> u,
         py::array_t<double> v)
 {
@@ -77,22 +78,22 @@ augment(py::array_t<double> cost_matrix,
     // u_data is the data from the numpy array u. i.e. u(i) is the i'th element.
 
     double minVal = 0;
-    int row_idx = cur_row;
-    int nr = cost_matrix.shape(0);
-    int nc = cost_matrix.shape(1);
+    int64_t row_idx = cur_row;
+    int64_t nr = cost_matrix.shape(0);
+    int64_t nc = cost_matrix.shape(1);
 
     // Crouse's pseudocode uses set complements to keep track of remaining
     // nodes.  Here we use a vector, as it is more efficient in C++.
-    int num_remaining = nc;
-    std::vector<int> remaining(nc);
-    for (int it = 0; it < nc; it++) {
+    int64_t num_remaining = nc;
+    std::vector<int64_t> remaining(nc);
+    for (int64_t it = 0; it < nc; it++) {
         // Filling this up in reverse order ensures that the solution of a
         // constant cost matrix is the identity matrix (c.f. #11602).
         remaining[it] = nc - it - 1;
         // remaining[it] = it;
     }
 
-    std::vector<int> path(nc, -1);
+    std::vector<int64_t> path(nc, -1);
     std::vector<double> shortestPathCosts(nc);
     std::fill(shortestPathCosts.begin(), shortestPathCosts.end(), INFINITY);
 
@@ -103,15 +104,15 @@ augment(py::array_t<double> cost_matrix,
     std::fill(SC.begin(), SC.end(), false);
 
     // find shortest augmenting path
-    int sink = -1;
+    int64_t sink = -1;
     while (sink == -1) {
 
-        int index = -1;
+        int64_t index = -1;
         double lowest = INFINITY;
         SR[row_idx] = true;
 
-        for (int it = 0; it < num_remaining; it++) {
-            int j = remaining[it];
+        for (int64_t it = 0; it < num_remaining; it++) {
+            int64_t j = remaining[it];
 
             double r = minVal + cost_data(row_idx, j)- u_data(row_idx) - v_data(j);
             if (r < shortestPathCosts[j]) {
@@ -130,7 +131,7 @@ augment(py::array_t<double> cost_matrix,
         }
 
         minVal = lowest;
-        int j = remaining[index];
+        int64_t j = remaining[index];
 
         // TODO: raise an exception if minVal is INFINITY
         if (minVal == INFINITY) { // infeasible cost matrix
@@ -150,7 +151,7 @@ augment(py::array_t<double> cost_matrix,
 
     // update dual variables
     // u_data(row_idx) += minVal;
-    for (int i = 0; i < nr; i++) {
+    for (int64_t i = 0; i < nr; i++) {
         if (SR[i]) {
             if (i == cur_row) {
                 u_data(i) += minVal;
@@ -161,14 +162,14 @@ augment(py::array_t<double> cost_matrix,
         }
     }
 
-    for (int j = 0; j < nc; j++) {
+    for (int64_t j = 0; j < nc; j++) {
         if (SC[j]) {
             v_data(j) -= minVal - shortestPathCosts[j];
         }
     }
 
     // augment previous solution
-    int col_idx = sink;
+    int64_t col_idx = sink;
     while (1) {
         row_idx = path[col_idx];
         row4col_data(col_idx) = row_idx;
@@ -192,22 +193,22 @@ py::array_t<T>  _fill(py::array_t<T> arr, T val) {
 py::tuple
 _solve(py::array_t<double> cost_matrix)
 {
-    int nr = cost_matrix.shape(0);
-    int nc = cost_matrix.shape(1);
+    int64_t nr = cost_matrix.shape(0);
+    int64_t nc = cost_matrix.shape(1);
 
     py::array_t<double> u = py::array_t<double>(nr);
     py::array_t<double> v = py::array_t<double>(nc);
-    py::array_t<int> row4col = py::array_t<int>(nc);
-    py::array_t<int> col4row = py::array_t<int>(nr);
+    py::array_t<int64_t> row4col = py::array_t<int64_t>(nc);
+    py::array_t<int64_t> col4row = py::array_t<int64_t>(nr);
 
     // TODO: Can we skip this or fold it into the allocation step above?
     _fill<double>(u, 0.);
     _fill<double>(v, 0.);
-    _fill<int>(row4col, -1);
-    _fill<int>(col4row, -1);
+    _fill<int64_t>(row4col, -1);
+    _fill<int64_t>(col4row, -1);
 
     // TODO: We only use cost_matrix through cost_matrix.unchecked<2>()
-    for (int cur_row = 0; cur_row < nr; cur_row++) {
+    for (int64_t cur_row = 0; cur_row < nr; cur_row++) {
         augment(cost_matrix, cur_row, row4col, col4row, u, v);
     }
